Signature processing fixes
It turns out we were reading the "V" component of the signature incorrectly.
We were looking at the first byte of a larger array, which was 0 when the total
value should have been 27. This happened to work out, because we were supposed
to be subtracting 27 from the provided value, but we weren't.