Ingest service
This is the initial work on the ingest service (minus the actual binary to
run it).

The handler injects a publisher and accounts service. The service accepts
either a 377 byte octet stream, or a JSON string up to 1KB. It parses
the data into an order object, then validates the signature and makes sure
that the order offers sufficient fees. If the account has been blacklisted,
the service silently drops the order. Otherwise the order is published via
the publishing service.