Add feeRecipient as a property of accounts
We had a new idea for an affiliates program, which would mean there
are multiple fee recipients. We'll store the multiple fee recipients
with the accounts service; it will just be an additional property of
an account and we'll look up the fee recipient in the account service
to determine elligibility.