Add mock channels and start of channels tests
This adds a "mock_channels" for in process message passing. It acts more like
a queue than a topic, in that each message gets relayed only once, and Acking /
Rejecting is fully implemented.

I've also implemented tests that run for both the mock channels, and if specified
via environment variables, the redis channesl.