Add relay commands
This adds a command for the fund verification relay, and one for the
simple relay. The simple relay takes messages from one source and
passes them to another. The fund verification relay is essentially
a simple relay with a filter that doesn't pass on messages if the
order doesn't have adequate funds.