Starting on monitoring service
The monitoring services are being written in javascript, due to
lack of support for necessary RPC functions in the go libraries.

So far I've written the base publishers, and a monitoring loop that
will maintain the block number so services can be resumed. Testing
is still needed, and we need to make sure that filters will be
automatically re-installed if the backend server changes.