Contracts for testing
OpenRelay doesn't directly depend on any original contracts, but
it does depend on having tokens to check the balances of, and a
0x exchange contract to monitor.

This adds a token factory for creating tokens contracts to serve as
the tokens we're checking for. It allows the owner of the contract to
issue themselves tokens any time they like, to simplify testing. I'm
not convinced it's a secure contract. It's based on the StandardToken
from the 0x contracts library, but I made it ownable and added issuance
for testing purposes.