Update relays to allow modification of values
I'm about to update the fundcheckrelay to include the MakerTokenAmountFilled
from the contract as it forwards messages to the next stage, but to do that
relays need to be able to support modifying the message they're relaying.

To do this, the filter now passes both the content (which can be just
delivery.Payload() if nothing is changing) as well as a boolean value indicating
wether to forward the value. If forwarding shouldn't happen, then the
string is likely empty.