Dynamo should only store order data in rawdata
The dynamodb data was storing the full 409 bytes for the order, now that
it was being provided to the indexer. This meant when it created the order
object from the bytes, it was overwriting the MakerTokenAmountFilled from
the one in the raw bytes, which wasn't getting updated.

The options were to either update the raw bytes as the MakerTokenAmountFilled
changes, which would have been complicated and require storing more data,
or only store the first 377 bytes of the raw data, which is easier and cheaper.