Add takerTokenAmountCancelled
To comply with the 0x standard API, we need to make a distinction between
takerTokenAmountFilled and takerTokenAmountCancelled. Previously I'd been
combining these into a single attribute, but then we wouldn't be able to
comply with the standard API (the api still isn't compliant, but the difference
is in organization of the output rather than captured data).

This commit adds tracking of the takerTokenAmountCancelled to the relevant services,
from the Golang data structure to the DynamoDB indexes.