FundChecker should account for filled / cancelled amounts
If the fund checker gets invoked on a partially filled or cancelled
order, it should account for that when validating the order. That is,
if half of the order has been filled, and the maker only has enough
to fill the other half of the order, the order is still valid.

While writing tests for this, I discovered some other issues with the
validator and the balance checker. Those issues have been fixed.