Prune old orders from order book
The queue orders script is now passing orders to the fill updater,
which passes any changes back to the indexer, and all orders off to
an inverse fund checker. If the inverse fund checker finds an order
with insufficient funds, it sends it back to the unindexer to be
removed from the index.

It also adds the fill updater before other fund checkers, so that
funds are only check with respect to the amount remaining.