Add fee info api
This adds a fee info API, compatible with the standard relayer api.

As the standard relayer API only allows us to specify a single valid
fee structure, we're assigning the entire fee to the maker. We will,
however, accept any combination of maker and taker fees that add up
to or exceed the amount specified as the maker fee in this response.