Require blockhash argument for search requests
If we include the blockhash in the search string, we can cache each
request very aggressively in CloudFront. Since orders are released
to the orderbook as blocks are mined and orders are filled when blocks
are mined, queries should never change between blocks. Thus the blockhash
can be used as part of the cache key, and the result can be cached
without needing to worry about cache invalidation, since the next block
will yield a new URL.