Update aws stack
This gets the aws swarm stack ready to deploy.

Part of this commit is a new ethnode container. It extends
ethereum/go-client, starting a light server with no RPC.
It checks periodically to see if the server has finished
syncing. If syncing is complete, it will enable the RPC server.

This ensures that until syncing is complete, the node will be excluded
from the haproxy round robin rotation.