The order validator should panic if eth-rpc is down
Per: https://github.com/NoteGio/openrelay/issues/17 we have an issue
where unavailability from the eth-rpc server causes all of the orders
to get dropped from the orderbook. This changes the order validator
to panic in the event that the RPC server is unavailable, or is
responding incorrectly. That panic should be caught at the relay,
the order returned to the source queue, then re-panicked.

Once the panic occurs the process will exit, and Docker will be responsible
for restarting the service. When the service starts, it will check
that the Ethereum server is responding correctly before taking messages
off the queue.