Create more log groups
Previously we had the worker nodes configured to log to AWS cloudwatch
logging groups based on the node they were on. This resulted in a
jumble of log streams that made it difficult to find what you were
looking for. Adding per-service log groups helps make logs easier
to sort through.