Remove volumes from eth nodes
Using the cloudstor volumes for the ethnodes turns out to be a
bit excessive. Moving a cloudstor volume between AZs takes a lot
of time, while rebuilding a light node from scratch is pretty quick
once you have the peers. It might make sense to go back to regular
volumes, but the cloudstor volumes seem to be more trouble than
they're worth.