Use filterProvider for blockMonitors
The providerEngine system offers a filterProvider which filters
blocks locally instead of installing filters on the RPC nodes. If
we use this for our monitors, we can simply load balance across a
pool of eth nodes instead of having to ensure that the monitoring
servers are always communicating with the same instance.

I still want to get the monitoring nodes using the websocket
subscriptions soon, but in the short term this should simplify a
lot of things.