Add postgres support via GORM
This is an initial commit, and postgres support is not comprehensive
or very well tested.

Per https://github.com/NoteGio/openrelay/issues/6 we're going to swap
out the DynamoDB packend with a Postgres Backend. I'm implementing that
change via GORM, which should give us compatibility with any of GORM's
other backend without much effort.

In order to manage how GORM interacts with the database, the Order
object now uses pointers to Address objects and Uint256 objects instead
of [20]byte and [32]byte arrays. This means orders must be initialized
to avoid nil pointer exceptions, but it allows us to have Addresses
and Uint256 objects implement driver.Valuer and sql.Scanner, so that
they can easily be persisted.