Add Indexer struct
This adds an indexer that takes in an order and inserts it into
the database, or updates the database if it already exists.

This can relace both the Indexer and FillIndexer from Python, as it
wil update or insert, instead of just doing one or the other.

Still needs to be wrapped in a channel consumer.