Create pg search service entrypoint
This adds the entrypoint for the search service. During testing,
I realized that I broke the ingest service earlier in the postgres conversin
when I changed ecSignature.V from a json.Number to a string. I changed
it back and fixed casting it as a json.Number elsewhere, so the issue
seems to have been resolved.