Separate blockhash logic into decorator
This separates out the blockhash redirection logic out into a decorator
that can wrap handler functions, so that it can be used with different
handler functions that need the same logic. It's similar to the way the
redirection was implemented in Python.

Also, fixed some issues with the tests, which were still looking for
ecSignature.V to be a string.