Add token_pairs endpoint
This adds the token_pairs endpoints.

This endpoint is intended to show what token pairs a relayer supports.
OpenRelay is intended to support arbitrary token pairs, so this API
will return any token pairs found in the system. Searching for token pairs
not found in the returned list will result in an empty list, but
it will still be possible to submit orders with token pairs missing from
this list.