Change password management for postgres services
I'm in the middle of cleaning up how passwords are handled for
postgres services. Previously, it would look for something in the
command line resembling a password, and failing to find one it would
try to load an environment variable.

I want to support AWS' parameter store, which is a fairly safe, auditable
way to store passwords in AWS, but I don't want to *just* support the AWS
parameter store.

The current approach adds a "common.GetSecret()" method, which takes the
following steps:

* If the value passed in on the command line is prefixed with file://,
  try loading the specified file. This is intended for Docker Swarm
  environments, which inject secrets into containers as files.
* If the value is prefixed with env://, try loading an environment variable.
* If the value is prefixed with ssm-param://, try loading from an AWS
  parameter store. Note that this requires credentials from the AWS credenital
  chain, and doesn't give the user a way to provide them directly.
* If the value isn't prefixed with any of the above, just use it as the password.

This will allow us to support additional secret loading methods in the future. Note
that a password or password lookup URI must always be specified on the command
line; even if you're using environment variables, you must tell on the command line
what environment variable to look in.

The automigrater has been made responsible for creating users and granting
permissions as it migrates the database. Right now, it's the only thing
using the credential lookup, but the other services that interact with postgres
will need to be updated to support it as well.