Get all postgres facing services using common.GetSecret()
This makes all postgres facing services use common.GetSecret() to retrieve
passwords. In production, this will retrieve passwords from ssm parameters,
but for lower environments it can pull from environment variables, files, or
just straight from the command line.

This commit also updates docker-compose-testrpc.yml to include docker image names,
as this is a convenient way to push up images, and I'm no longer maintaining the
AWS version since we're not using a swarm anymore.

Finally, it adds a certificate file to docker images that interface with ssm, as
those need to verify certificates, which requires having a certificate file.

At this point everything is ready to go, and is actually deployed to production.
The only thing not currently working is a healthcheck for the searchapi.
Since all of the members of the group are failing healthchecks, the load
balancer is still passing requests through, so it's usable (but I'm not
really advertising it yet). Tomorrow I plan to implement a health check
view, which will return the current blocknumber and count of the orders
in the system, just to confirm redis and postgres communication are working.