Add exchange splitter
The exchange splitter is intended to be a stage in the pipeline that
will split messages to different pipelines based on the exchange
address. This way we can have different pipelines for different exchanges,
and exchanges are currently our best indicator of network. In the short
term, this will let us have one pipeline for mainnet and one pipeline for
ropsten, which will fork right after the ingest stage and right after
the ordercheckenqueue stage.