Use RPC Based TokenProxy and ZRX Token lookups
As we move to support testnet, we'll have to deal with more than
one Exchange contract and mor ethan one ZRX Token contract on the
different networks. Rather than continuing to keep this information
in Redis where we'd have to keep it up to date and refactor to support
multiple exchanges / token addresses / token addresses, the services
that need that information can pull it from the RPC server based on
the order's exchange address.

This doesn't quite cover everything, as the ingest service is still
pulling the ZRX price from redis, but at least for now I'm okay saying
that the mainnet and testnet base fees will be the same.