Add SpendRecorder
This adds the bulk of a SpendRecorder (still need to make a command
and container for it). It watches a channel for messages indicating
that a token has been transferred, and updates any orders that would
become unfillable as a result of the transfer.

Still need to make a command and container to reflect this functionality,
as well as a service to watch the chain and produce the incoming
messages.