Add Golang Block Monitor Service
This implements a block monitor service in Go, which works somewhat
differently from the javascript predecessor.

At a configurable interval, it queries the RPC server for $LATEST_BLOCK + 1
where $LATEST_BLOCK is the last block it has processed. It makes sure that
it has seen the parent of the current block, then emits the:

* Block Number
* Block Hash
* Block Bloom Filter

to a target channel. If the block monitor has not seen the parent block,
this indicates a chain reorg, so it will seek back to find a common
ancestor between a previously emitted block and the newly discovered block.
Once it finds the common ancestor, it will emit each block from the new chain.

The goal of this services is that other monitoring services will act as consumers
of the block monitoring service. They will consume blocks from the block monitor,
check their filters against the bloom filter passed by the block monitor, and if
necessary query RPC for pertinent events. This should cut back on the load on
the RPC server quite a bit. With the javascript implementation every monitoring
service queried the RPC server to find new blocks. With this implementation, only
the block monitor will query the RPC server to discover new blocks, and other
monitors will only query the RPC server when the block bloom filter indicates a
relevant event. For some monitoring services, it's likely that there will be
relevant events every block, but the block monitor still keeps those services
from having to poll to discover blocks.