Update order validator
The order validator hadn't been updated since it used to get filled
quantities at the same time. In the microservice those functions got
separated out, but for a single validator command it's still useful to
have them in one place.

This came up as the result of some troubleshooting, and the order
validator was insufficient.