Add allowance check consumer
This adds an allowance checker downstream from the block monitor. When
it gets a message about a block it checks the bloom filter for any allowances
set on the TokenTransferProxy. If it finds a match, it queries for logs about
matches. For any match it finds, it emits a SpendRecord event on a publisher
channel.

While allowances aren't actually spends, they have the same downstream effect.
If we see an allowance lowered beyond a certain point it has the same effect
on the fillability of an order as if the same account had spent down to the
same level. Thus at this point, we don't need to differentiate events for
spends vs allowance changes.