Remove unneeded services
This commit removes the services that will no longer be needed as we
deprecate the DynamoDB storage engine and move to the new system of
monitors.

This completely eliminates the Python codebase, and eliminates the
Node codebase from production, though Node is still used for initializing
the test network, at least for now.