Cache balances in the fund checker
In the upcoming embiggen exercise, we will be loading millions of orders
signed by a small set of makers for a single token. Thousands of these orders
will be loaded every block, and there is no chance that a maker's balance
will change between blocks. As such, we can cache balances so long as we
invalidate the cache every time a new block is mined.

Whenever a message comes in on a channel, the cache is reset to an empty map.
The intention is that this channel will be a topic that gets signals every time
a new block is found. When a user's balance or allowance is checked, the result
will be stored in the cache, and it won't need to be rechecked until the next
block.