Add bloom filter to fill updater
Since the vast majority of orders won't be filled or cancelled,
a bloom filter is a cheap way to avoid RPC calls. This configures
the fill monitor to manage a bloom filter for resumption, and the
fill updaters to maintain a bloom filter from the fill logs.

If an order is not in the bloom filter, we don't bother making the
RPC calls to get fill and cancellation amounts.