Fix bugs in accounts / affiliate storage
This fixes an issue with the storage of accounts and affiliates.
Both were supposed to be JSON serialized, but because the variables
were private the JSON encoder skipped over them.

I also made the keys for affiliates and accounts use hex encoded
account addresses instead of just passing the raw bytes. Redis is
okay with raw bytes, but it's harder to interact with and inspect
manually.