Limit expiration date range
This limits expiration dates to greater than now, but less than 1,000
years in the future.

I had to update tests so that the time would be in this range. That means
tests will start failing on September 22, 2153. If OpenRelay is still in
use in 135 years, I apologise to whoever had to troubleshoot this one.