Streamlining / Concurrency
As we start ingesting the Embiggen orders, we're finding bottlenecks
in the ingest process. This streamlines the fund checker, skipping lookups
where no value is required (this will skip fee lookups when there is no
makerfee).

It also adds concurrency to relays, managed by a semaphore. Right now,
most relays use a CONCURRENCY environment variable to determine the level of
concurrency, defaulting to five goroutines. The exception is the delay relay,
which only supports a single goroutine (but fortunately, the delay relay
doesn't seem to be any more of a bottleneck than it's supposed to be).