Only flush once at a time
The DelayRelay has been ending up with a lot of extra sentinel messages
on the queue. I believe this happens because the topic consumer happens
concurrently, while the message consumer happens sequentially, so if the
consumer isn't waiting on delayChan when a Flush() happens we end up with
multiple sentinel messages on the queue at the same time.

Adding a sync.Once to the DelayRelay can ensure that only one signal message
can cause a flush at a time, publishing only one message and then publishing
to the delayChan. If other flushes try to happen at the same time, they will
get skipped. After blocking on sending to delayChan, relay.Once is replaced so
that the next signal will happen again.