Efficiency improvement in spend monitor
The vast majority of token spends we monitor for will have nothing
to do with the 0x exchange. Those tokens will have no allowance on
the 0x token proxy, and thus the allowance will be 0. If the allowance
is 0, the balance in our spend record will be 0, so we don't need
to get the actual balance. This should cut the RPC calls made by the
spend monitor by slightly less than half.