Add affiliate monitoring microservice
This adds a microservice to monitor the blockchain for affiliate
registrations, and adds them to the affiliate service so we'll
accept them as valid fee recipients.

This is based on a yet-to-be-released Affiliate and AffiliateFactory
contract, which will allow people to sign up as an affiliate. Signing
up through the AffiliateFactory will create an Affiliate contract
specific to that affiliate. OpenRelay will monitor the AffiliateFactory
for new Affiliate contracts, and will accept them as fee recipients.

The docker-compose-testrpc.yml is currently configured to watch an address
where the AffiliateFactory contract will end up if you run
"truffle deploy --network=testrpc" from the affiliate contract git repository,
but that has not been published at the time of this commit, and the
contract is not deployed to the testrpc network by default.