Don't support ERC721 just yet
For our initial Kovan release we're just going to support ERC20, and
ERC 721 support will come soon after. There are some minor additions
that will be required in the fund checker, spend monitor, and allowance
monitor, so initially we're going to support ERC20, but expect to have
the rest ironed out very soon.