Add cancellation conumser
This adds a cancellation consumer. To make the cancellation consumer
work, we needed to be able to unmarshal Cancellations. It seemed
that the easiest way to do this was to make Addresses and Uint256s
Unmarshallable.

When Order was originally implemented, it used [20]byte and [32]byte,
and only later did we add the Address and Uint256 types. That made it
necessary to implement the jsonOrder class to unmarshal Orders. Now that
Addresses and Uin256s are individually unmarshallable, we could probably
clean up Order significantly by removing the jsonOrder class and marshalling
stuff.