Change multistage build approach
We're working on moving Docker builds into circleci, and want to
parallelize the builds of the binaries. The current approach builds
all of the binaries at once, which precludes building them in parallel.

The idea here is that we will tag a local "corebuild" image that handles
the build. For docker compose builds, we will use Dockerfile.composebase,
which will build all of the binaries up front. For circle ci builds, we will
use Dockerfile.cibase, which will copy the binaries in from the local system
after they were built in parallel.