Moved some code from monitor.js to blockMonitor.js
Edited blockMonitor.js to take on more of the responsibility of
formatting the data before passing it to monitor.js. That way
blockMonitor.js can be built to deal with different types of data
without changing monitor.js.

I have NOT fully tested this change and may need further review.