Clean up Kovan Block Monitor
There were a couple of issues with TechAdvancer's update to the
block monitor - primarily use of await, which can't be used because
the parent function wasn't defined as async. I just rearranged things
a little bit to deal with the async stuff inside the promise, instead
of relying on the async keywords.

I also added a Dockerfile so we can run it.