Better event validation
Until now we had assumed that events that matched on topic[0] would
otherwise comply with the event signature. It turns out this is not
always true, so we need additional validation.

We probably ought to add validation to some of the other monitors
soon, but this covers the performance critical ones, so it shouldn't
be a huge deal if the others crash and get restarted for the time being.