SRA v2 implementation
This is an almost-complete implementation of the 0x SRA v2.
I need to implement the fee_recipients API, and double check
everything else against the spec, but this should be very close
to compliant.

The biggest change is to support the networkId parameter. For that,
we had to have a mapping of network IDs to exchanges in the database.
I'm currently having the application look up the network id -> exchanges
list once, then cache it in memory. It *could* be done with joins
in the queries, but that would be more complicated than I wanted to try
and get GORM to do, and since the number of exchanges will never be
enormous, caching them all in memory should give us a slight performance
bump.

Now that we have an exchange -> network mapping, we should consider refactoring
a number of the other services that currently take an exchange address, and
refactor them to take a network ID (or just get the network ID from their
RPC interface, if applicable), and enable those processes to support multiple
concurrent exchanges so long as they use the same network.