Prepare for v2 mainnet launch
This modifies the exchange splitter so that instead of having a defined
mapping, of exchange -> queue, it will derive the queue name from the
exchange address. This will let us separate a shared stack that includes:

* Ingest
* Search
* Indexing
* Exchange Splitter

From exchange-specific validation stacks. The exchange splitter will
output to a queue based on the exchange address, and the validation
stacks will need to know to pull from there.

This also adds the v2 mainnet exchange to the ingest service.