Fix bug in bloom filter testing
It turns out that Bloom.Test() doesn't actually work with Ethereum
bloom filters if any of the topics have leading 0 bytes. That issue
is documented here: https://github.com/ethereum/go-ethereum/issues/18092

This switches to using the BloomLookup() method, which works correctly.

The only place this was actually a problem was the erc721approvals and
approvals monitors. All of the others only relied on topics that were
32 bytes instead of 20 or less.