Validate terms acceptance on ingest
This ensures that the maker of an order has accepted the terms of
use before putting the order in the order book. This is the first
time the ingest process has had access to the database (and the command
utility has not been updated to reflect this, so right now builds will fail).
Since it now has access to the database, we're also pulling valid exchanges
from the database instead of hardcoding.

Both the ToS and Exchange lookup are read-only (and should use a read-only
database connection), as well as asynchronous (they initiate the lookup then
do a bunch of other validation before getting the results from the database).
Exchange lookups are cached heavily, and shouldn't lead to db lookups often.