Add cache control to errors on TOS
We found that when dApps query for whether an account is authorized,
cloudfront caches the answer. After they authorize, the dApp asks again,
and cloudfront returns the cached "no they don't exist" even though
the authorization changed the correct answer.

We're making error pages uncacheable here, which we should consider
doing for other endpoints as well.