Add contract filter support to pools
This adds the ability to the pools class to filter orders based
on an on-chain contract. The order gets ABIv2 encoded and sent to
the contract, which will return true or false to indicate whether
the order should be added to the orderbook. Since the abigen tool
doesn't support ABIv2, we had to do the encoding manually, which is
tedious, but seems to be working in tests.

The next step will be to create a microservice to check orders against
pools as they pass through.