Add pool filter pipe stage
This adds a command line tool that filters orders based on the pool's
specified filtering contract. We had to update pools to specify filters
and sender addresses based on the target network, otherwise we had to
validate orders against a different chain than the order would be filled
on in some circumstances, which doesn't really make sense.

Aside from a bit of debugging, this should finish out the minimal implemenation
of pools.