Websockets service
This adds a websockets service, which accepts subscriptions over websockets
and forwards relevant orders to subscribers.

It is broken up into `channels/ws`, which offers the channels interface for
websockets connections, and the `subscriptions` package, which listens for
subscriptions on one channel and publishes relevant orders to a publisher
channel. The subscriptions package is applicable to channels more generally,
and could potentially be used for order routing independent of websockets.