Docker Compose Cleanup
This cleans up the docker-compose file a bit, but shouldn't change
default behaviors.

It makes sure that everywhere Redis, Postgres, or an Ethereum Client is
needed, they can be pointed to external systems by setting the
REDIS_HOST, POSTGRES_HOST, and ETHEREUM_NODE environment variables
respectively. This makes it possible to run in a more production-like
settings off of the same docker-compose.yml