Publish relay messages in goroutines
Occasionally, for reasons I have yet to determine, messages hang
when we attempt to publish them. When that happens, none of the
subsequent publishers from the relay get the message, and the
message never gets acknowledged.

By firing them off as goroutines, we will potentially have a goroutine
hanging around (which we'll still need to figure out), but at least
we won't miss the publishers after the one that hangs.