Add admin endpoint, pool order cancellation
In this commit we add an admin interface for pools with a cancellation
method. Each pool will have an admin key, which may be used with a
sha256 HMAC to cancel orders. Only orders submitted through that pool
may be canceled. Each request must also specify an expiration time,
which can be no more than zero seconds in the past or ten seconds in
the future (to prevent replay attacks).

Additionally we may add more admin functions, but order cancellation
is the immediate requirement.